+++ mars/tools/MemoryReferenceVisualization.java
@@ -670,10 +670,10 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
              public void stateChanged(ChangeEvent e) { 
                JSlider source = (JSlider)e.getSource(); 
                if (!source.getValueIsAdjusting()) { 
-                  counterIndex = (int)source.getValue();
+                  counterIndex = source.getValue();
                } 
                else {
-                  int count = countTable[(int)source.getValue()];
+                  int count = countTable[source.getValue()];
                   sliderLabel.setText(setLabel(count));
                   currentColorButton.setBackground(counterColorScale.getColor(count));
                }
