+++ mars/venus/RunSpeedPanel.java
@@ -139,7 +139,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
           public void stateChanged(ChangeEvent e) { 
             JSlider source = (JSlider)e.getSource(); 
             if (!source.getValueIsAdjusting()) { 
-               runSpeedIndex = (int)source.getValue();
+               runSpeedIndex = source.getValue();
             } 
             else {
                sliderLabel.setText(setLabel(source.getValue()));
